move configure methods back to base action class
